backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Flask-Limiter==3.5.0
structlog==23.2.0
PyMySQL==1.1.0
gunicorn==21.2.0
Flask-Migrate==4.0.5
